use version-dependent sources (Scala 2 vs Scala 3)
for now, Scala 3 uses stubs
