+++ compiler/src/main/scala-3/play/twirl/compiler/FunctionCompiler.scala
@@ -1,0 +1,9 @@
+/*
+ * Copyright (C) Lightbend Inc. <https://www.lightbend.com>
+ */
+package play.twirl.compiler
+
+object FunctionCompiler {
+  def getFunctionMapping(signature: String, returnType: String): (String, String, String) =
+    ???
+}
